Add device auto-discovery and cli option
Add `--device` cli option to specify which CI device is
being tested.  If the cli option is not specified, then
we fallback to the usual GAUDI2_CI environment variable.
Finally, if neither is specified, then try to get the
device name automatically from the habana frameworks
API.

This setting is to prepare for gaudi3 specific (and
future devices) test cases and baseline references.

Eventually, we can deprecate the GAUDI2_CI environment
variable once the rest of the test framework is
migrated to the new device declaration.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>