Silence `Trainer.tokenizer` warnings
Recent versions of `transformers` deprecated the usage of
`Trainer.tokenizer` in favor of `Trainer.processing_class`. However, the
property method of the tokenizer attribute triggers a `logging.warning`
each time the property is accessed. This causes the script output to be
flooded with these tokenizer warnings. Until the next `transformers`
upgrade, we want to silence these warnings.