Synchronize processes when saving model
When running a script on multiple HPUs using MPI, the script hangs.
An investigation releaved that the problem is related to a `barrier`
in the `Trainer` code [1], i.e., not all processes reach this barrier.
In fact, the `self.state.best_model_checkpoint` is only set in certain
process, whereas in other ones the variable is set to `None`. The
variable should be set by the `_save_checkpoint` method [2] (which is
in turn called by the `_maybe_log_save_evaluate` method [3]). Debugging
showed that when different processes call the `os.path.exists(best_checkpoint_dir)`
function [4], they get random results (sometimes it returns true, always
a different subset of processes). This shows that there is a race
condition on this directory check - i.e., the main process (rank zero)
writes the model weights, but before it finishes other ranks try to
check for the existence of this directory. This commit introduces
a synchronization barrier before checking if the directory exists.

[1] https://github.com/huggingface/optimum-habana/blob/v1.19-release/optimum/habana/transformers/trainer.py#L1144
[2] https://github.com/huggingface/transformers/blob/v4.51-release/src/transformers/trainer.py#L3187
[3] https://github.com/huggingface/optimum-habana/blob/v1.19-release/optimum/habana/transformers/trainer.py#L1296
[4] https://github.com/huggingface/transformers/blob/v4.51-release/src/transformers/trainer.py#L3206