Disable image processor loading for `THUDM/glm-4v-9b`
After the Transformers 4.51->4.55 upgrade, the `pipeline(...)` function
was completely refactored [1]. The logic related to loading individual
components of a given pipeline was also updated. Now, despite setting
the `image_processor=None` argument, when dealing with the `chatglm`
model, the code in the `pipeline(...)` will still try to load the image
processor, which fails eventually (`chatglm` does not have an image
processor class). To fix that, this commit overrides the
`ImageToTextPipeline._load_image_processor` property to `False`.

[1] https://github.com/huggingface/transformers/commit/3d8be20cd2f76aa03d3f42808f2c9b36c94608b3