GaudiTrainer: Take accuracy measurement outside of timer
In GaudiTrainer, accuracy calculation happened between start_time
and end_time, so its duration was included in calculating time for
performance measurements.
It was negligible for datasets < 4.0.0, but with the new package,
it takes much more time due to numpy array conversion deep in
the scikit-learn library, causing incorrect perf reporting.

Signed-off-by: Urszula <urszula.golowicz@intel.com>