Fix Mllama attention selection (#2343)
There was a wrong assumption in PR [1] that the `vision_model` attention
implementation should always use the `sdpa` option. However, if the
flag for `FusedSDPA` was passed (i.e., `use_flash_attention`), the model
should not use `sdpa` but `FusedSDPA` in the vision model. This commit
fixes the selection of the attention implementation.

[1] https://github.com/huggingface/optimum-habana/pull/2319